requirements.txt
quart
quart-cors
hypercorn
python-dotenv
google-generativeai
langchain-mcp-adapters
anthropic
boto3
openai